+++ services/eip/v2/EipRegion.ts
@@ -17,6 +17,7 @@ export class EipRegion {
     public static CN_NORTH_2 = new Region("cn-north-2", "https://vpc.cn-north-2.myhuaweicloud.com");
     public static LA_SOUTH_2 = new Region("la-south-2", "https://vpc.la-south-2.myhuaweicloud.com");
     public static NA_MEXICO_1 = new Region("na-mexico-1", "https://vpc.na-mexico-1.myhuaweicloud.com");
+    public static AP_SOUTHEAST_4 = new Region("ap-southeast-4", "https://vpc.ap-southeast-4.myhuaweicloud.com");
     
 
     private static STATIC_FIELDS = EipRegion.createStaticFields();
@@ -38,6 +39,7 @@ export class EipRegion {
         map.set("cn-north-2", EipRegion.CN_NORTH_2);
         map.set("la-south-2", EipRegion.LA_SOUTH_2);
         map.set("na-mexico-1", EipRegion.NA_MEXICO_1);
+        map.set("ap-southeast-4", EipRegion.AP_SOUTHEAST_4);
         return map;
     }
     public static valueOf(regionId: String) {