+++ services/eip/v2/model/ChangePublicipToPeriodRequest.ts
@@ -1,0 +1,12 @@
+import { ChangeToPeriodReq } from './ChangeToPeriodReq';
+
+
+export class ChangePublicipToPeriodRequest {
+    public body?: ChangeToPeriodReq;
+    public constructor() { 
+    }
+    public withBody(body: ChangeToPeriodReq): ChangePublicipToPeriodRequest {
+        this['body'] = body;
+        return this;
+    }
+}