+++ services/functiongraph/v2/model/ListReservedInstanceConfigsRequest.ts
@@ -2,6 +2,8 @@
 
 export class ListReservedInstanceConfigsRequest {
     private 'function_urn'?: string | undefined;
+    public marker?: string;
+    public limit?: string;
     public constructor() { 
     }
     public withFunctionUrn(functionUrn: string): ListReservedInstanceConfigsRequest {
@@ -13,5 +15,13 @@ export class ListReservedInstanceConfigsRequest {
     }
     public get functionUrn() {
         return this['function_urn'];
+    }
+    public withMarker(marker: string): ListReservedInstanceConfigsRequest {
+        this['marker'] = marker;
+        return this;
+    }
+    public withLimit(limit: string): ListReservedInstanceConfigsRequest {
+        this['limit'] = limit;
+        return this;
     }
 }