+++ services/functiongraph/v2/model/ListReservedInstanceConfigsResponse.ts
@@ -4,15 +4,21 @@ import { ReservedInstanceConfigs } from './ReservedInstanceConfigs';
 import { SdkResponse } from "@huaweicloud/huaweicloud-sdk-core/SdkResponse";
 
 export class ListReservedInstanceConfigsResponse extends SdkResponse {
-    public reservedinstances?: Array<ReservedInstanceConfigs>;
+    private 'reserved_instances'?: Array<ReservedInstanceConfigs> | undefined;
     private 'page_info'?: PageInfo | undefined;
     public count?: number;
     public constructor() { 
         super();
     }
-    public withReservedinstances(reservedinstances: Array<ReservedInstanceConfigs>): ListReservedInstanceConfigsResponse {
-        this['reservedinstances'] = reservedinstances;
+    public withReservedInstances(reservedInstances: Array<ReservedInstanceConfigs>): ListReservedInstanceConfigsResponse {
+        this['reserved_instances'] = reservedInstances;
         return this;
+    }
+    public set reservedInstances(reservedInstances: Array<ReservedInstanceConfigs> | undefined) {
+        this['reserved_instances'] = reservedInstances;
+    }
+    public get reservedInstances() {
+        return this['reserved_instances'];
     }
     public withPageInfo(pageInfo: PageInfo): ListReservedInstanceConfigsResponse {
         this['page_info'] = pageInfo;